+++ noodles-fasta/src/repository/adapter.rs
@@ -5,5 +5,5 @@ use crate::Record;
 /// A repository adapter.
 pub trait Adapter {
     /// Returns the record with the given name.
-    fn get(&mut self, name: &str) -> Option<io::Result<Record>>;
+    fn get(&mut self, name: &[u8]) -> Option<io::Result<Record>>;
 }